pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "3.11.4"
asyncpg= "^0.27.0"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
fastapi = "^0.97.0"
SQLAlchemy = "^2.0.16"
alembic = "^1.11.1"
uvicorn = "^0.22.0"
pydantic = "^1.10.9"
pillow = "^9.4.0"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
python-dateutil = "^2.8.2"
celery = "^5.3.1"
flower = "^2.0.0"
redis = "^4.6.0"
httpx = "^0.24.1"
email-validator = "2.0.0.post2"
python-multipart = "^0.0.6"
pyjwt = "^2.8.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
flake8-bugbear = "^23.3.12"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-quotes = "^3.3.2"
flake8-todo = "^0.7"
flake8-variables-names = "^0.0.5"
autoflake = "^1.3.1"
[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

